sistemata organizzazione packages e sistemata relazione tra ordine e prodotto, l'app spring non parte, da sistemare

+++ src/main/java/org/datotoweb/models/Ordine.java
@@ -4,6 +4,7 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.util.Date;
 import java.util.LinkedHashSet;
 import java.util.Set;
 
@@ -23,7 +24,7 @@ public class Ordine
     private Utente utente;
 
     @Column(name = "prezzo", nullable = false)
-    private double prezzo;
+    private float prezzo;
 
     @Column(name = "stato", nullable = false)
     private String stato;
